chore: remove post-process
